redeployment 5

+++ inngest/functions.js
@@ -1,14 +1,11 @@
 import { inngest } from "./client";
 import { prisma } from "@/lib/prisma";
-//inngest Function to save user data to a database
+
+// Sync user creation
 export const syncUserCreation = inngest.createFunction(
   {
     id: "sync-user-create",
-    triggers: [
-      {
-        event: "clerk/user.created",
-      },
-    ],
+    triggers: [{ event: "clerk/user.created" }],
   },
   async ({ event }) => {
     const { data } = event;
@@ -17,40 +14,50 @@ export const syncUserCreation = inngest.createFunction(
       data: {
         id: data.id,
         email: data.email_addresses[0].email_address,
-        name: `${data.first_name} ${data.last_name}`,
+        name: `${data.first_name ?? ""} ${data.last_name ?? ""}`.trim(),
         image: data.image_url,
       },
     });
   }
 );
 
-
-//Inngest Function to update user data in database 
-
+// Sync user update
 export const synUserUpdation = inngest.createFunction(
-    {id: 'sync-user-update'},
-    {event: 'clerk/user.update'},
-    async({ event }) => {
-        const {data} = event
-        await prisma.user.update({
-            where:{id:data.id},
-            data:{
-                email:data.email_addresses[0].email_address,
-                name:`${data.first_name} ${data.last_name}`,
-                image:data.image_url,
-            }
+  {
+    id: "sync-user-update",
+    triggers: [{ event: "clerk/user.updated" }],
+  },
+  async ({ event }) => {
+    const { data } = event;
 
-        })
-    }
-)
-//Inngest Function to delete user from database
+    await prisma.user.update({
+      where: {
+        id: data.id,
+      },
+      data: {
+        email: data.email_addresses[0].email_address,
+        name: `${data.first_name ?? ""} ${data.last_name ?? ""}`.trim(),
+        image: data.image_url,
+      },
+    });
+  }
+);
+
+// Sync user deletion
 export const syncUserDeletion = inngest.createFunction(
-    {id: 'sync-user-delete'},
-    {event: 'clerk/user.deleted'},
-    async({ event }) => {
-        const {data} = event
-        await prisma.user.delete({
-           where:{id: data.id}
-        })
-    }
-)
+  {
+    id: "sync-user-delete",
+    triggers: [{ event: "clerk/user.deleted" }],
+  },
+  async ({ event }) => {
+    const { data } = event;
+
+    if (!data?.id) return;
+
+    await prisma.user.delete({
+      where: {
+        id: data.id,
+      },
+    });
+  }
+);
